Restructure packaging for online banking jp
COAND-497
